v3
- implemented the new steam-api: tweets are now created even faster!
inventories are parsed every 15 seconds now.

(note: they changed the rate limits for the old method to 2 requests per
MINUTE. (wtf? before it was 15 per minute, which was already pretty low,
but this?) the new api is much more organized and the rate limits are
way higher with around 500 requests per minute.)